+++ languages/de_DE/messages.php
@@ -91,4 +91,7 @@ return [
     'Error while uploading images'                                       => 'Fehler beim hochladen der Bilder',
     'Thumbnails rebuilded. Please reload the page to see changes!'       => 'Thumbnails neu erstellt. Bitte Seite aktualisieren um die Änderungen zu sehen!',
     'Rebuild thumbnails'                                                 => 'Thumbnails neu erstellen',
+    'Rotate image'                                                       => 'Bild drehen',
+    'Cache cleared.'                                                     => 'Cache gelöscht.',
+    'Please choose a gallery first'                                      => 'Bitte wähle zuerst eine Galerie aus',
 ];