+++ src/Controller/ImageApiController.php
@@ -2,6 +2,7 @@
 
 namespace Shw\Gallery\Controller;
 
+use Gregwar\Image\Image as GImage;
 use Pagekit\Application as App;
 use Shw\Gallery\Model\Image;
 
@@ -32,8 +33,15 @@ class ImageApiController
         }
 
         unset($data['modified']);
-
         $image->save($data);
+
+        if ($data['rotate']) {
+            $path = 'storage/shw-gallery/'.$data['filename'];
+
+            $img = GImage::open($path)
+                ->rotate(-$data['rotate'])
+                ->save($path, (int) App::module('gallery')->config('images.image_quality'));
+        }
 
         return ['message' => 'success'];
     }